Move English Wiktionary thesaurus extract function to the `extractor` folder
and load thesaurus extractor function according to the dump file
language code with `importlib.import_module()`.
